[iter15] updates after review
