+++ internal/server/storage/factory.go
@@ -2,11 +2,11 @@ package storage
 
 import (
 	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/mobypolo/ya-41go/cmd"
+	"github.com/mobypolo/ya-41go/internal/server/config"
 	"time"
 )
 
-func MakeStorage(cfg cmd.Config, db *pgxpool.Pool) *PersistentStorage {
+func MakeStorage(cfg config.Config, db *pgxpool.Pool) *PersistentStorage {
 	if cfg.DatabaseDSN != "" && db != nil {
 		return NewPersistentStorageWithPostgres(db)
 	}
